Feat: Добавлен компонент MoviesList
